legacy-data-manager/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
langchain==0.0.350
openai==1.3.5
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2 